Fix resizing windows!!!!
This took 5 days...

The final hint was found here: https://www.gamedev.net/forums/topic/704005-resize-window-in-d3d12/5413557/, where the OP signalled the graphics queue and then waited on it in order to flush the queue.

This commit uses the main branch of windows-rs to use the new `windows::core::ManuallyDrop` type. We can go back to using regular versions once windows-rs gets a new version.